inference/web_app/requirements.txt
google-generativeai
python-dotenv
flask
numpy
opencv-python
pillow
torch
transformers
